Fix module injection for `tree` and `cipher` modules
See: https://rust-lang.github.io/rust-clippy/master/index.html#module_inception
